% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_data.R
\name{attach_data}
\alias{attach_data}
\title{attach_data
Function to set up the data for import into stan based on the model and data
frame defined by the user.}
\usage{
attach_data(
  model,
  dat,
  time_var = "dt",
  count_vars = NULL,
  prevcount_vars = NULL,
  predictor_vars = NULL,
  hierarchical_group = NULL
)
}
\arguments{
\item{model}{bp_model defined by the user}

\item{dat}{data frame that contains counts for current and previous time and
predictor variables}

\item{time_var}{variable name used to designate the time}

\item{count_vars}{character vector of column names in dat that are the
current time object counts}

\item{prevcount_vars}{character vector of column names in dat that are
the previous time object counts}

\item{predictor_vars}{character of variable name in dat that is the predictor
variable for the statistical model}

\item{hierarchical_group}{character of variable name representing the group
variable if the model has random effects}
}
\description{
attach_data
Function to set up the data for import into stan based on the model and data
frame defined by the user.
}
