% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{build_parents_and_eventmatrix}
\alias{build_parents_and_eventmatrix}
\title{Function to build the parants and list of events into a vector and matrix}
\usage{
build_parents_and_eventmatrix(model)
}
\arguments{
\item{model}{bpmodel structure}
}
\description{
Function to build the parants and list of events into a vector and matrix
}
