% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_stancode.R
\name{create_stan_code}
\alias{create_stan_code}
\title{Generate Stan code from a provided bp model}
\usage{
create_stan_code(model)
}
\arguments{
\item{model}{A bp model that has been processed and parsed}
}
\description{
Generate Stan code from a provided bp model
}
