% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_stancode.R
\name{create_stan_error_block}
\alias{create_stan_error_block}
\title{Create Stan observed error chunk for model from parsed text}
\usage{
create_stan_error_block(model)
}
\arguments{
\item{model}{A bp model that has been processed and parsed}
}
\description{
Create Stan observed error chunk for model from parsed text
}
