% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_stancode.R
\name{create_transparameter_block}
\alias{create_transparameter_block}
\title{Create Stan transformed parameter block from parsed text}
\usage{
create_transparameter_block(model)
}
\arguments{
\item{model}{A bp model that has been processed and parsed}
}
\description{
Create Stan transformed parameter block from parsed text
}
