% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{format_constraint}
\alias{format_constraint}
\title{Format the constraints from a processed model}
\usage{
format_constraint(constraint)
}
\arguments{
\item{constraint}{a constraint on a parameter or data variable}
}
\description{
Format the constraints from a processed model
}
