% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{merge_constraints}
\alias{merge_constraints}
\title{Function to merge all constraints from the blocks of the model definition}
\usage{
merge_constraints(params_from_constraints, params_from_model)
}
\arguments{
\item{params_from_constraints}{all parameters pulled from the constraints arguments}

\item{params_from_model}{all parameters pulled from the model arguments}
}
\description{
Function to merge all constraints from the blocks of the model definition
}
