% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model.R
\name{new_transition}
\alias{new_transition}
\title{new_transition
constructor for object of class \code{transition}}
\usage{
new_transition(name, parent, offspring, model)
}
\arguments{
\item{name}{name for the transition event}

\item{parent}{the parent type for the transition}

\item{offspring}{the update to the system after the transition}

\item{model}{the user-defined bp_model output of process_model}
}
\value{
A new branching process rate object
}
\description{
new_transition
constructor for object of class \code{transition}
}
