% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parse_formulas}
\alias{parse_formulas}
\title{Function to parse all formulas from a model}
\usage{
parse_formulas(model)
}
\arguments{
\item{model}{formula block}
}
\description{
Function to parse all formulas from a model
}
