% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parse_model}
\alias{parse_model}
\title{Function to parse through the model and get the required pieces to construct
the stan code}
\usage{
parse_model(model)
}
\arguments{
\item{model}{model structure}
}
\description{
Function to parse through the model and get the required pieces to construct
the stan code
}
