% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parse_parameters}
\alias{parse_parameters}
\title{Function to parse parameters and separate name and constraint.
Requirement here to include all parameters and any possible constraints}
\usage{
parse_parameters(parameters)
}
\arguments{
\item{parameters}{parameters block from a model}
}
\description{
Function to parse parameters and separate name and constraint.
Requirement here to include all parameters and any possible constraints
}
