% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parse_priors}
\alias{parse_priors}
\title{Function to parse formula and extract parts}
\usage{
parse_priors(formulas, hierarchical)
}
\arguments{
\item{formulas}{formulas from a model}

\item{hierarchical}{hierarchical parameters from a model}
}
\description{
Function to parse formula and extract parts
}
