% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parse_single_formula}
\alias{parse_single_formula}
\title{Function to parse a single formula formula and extract parts}
\usage{
parse_single_formula(formula, predictor_names)
}
\arguments{
\item{formula}{formula block}

\item{predictor_names}{names of the predictor values (concentration, etc)}
}
\description{
Function to parse a single formula formula and extract parts
}
