% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{parse_single_formula_nomodel}
\alias{parse_single_formula_nomodel}
\title{Function to parse formula when there is no model}
\usage{
parse_single_formula_nomodel(formula, rateName)
}
\arguments{
\item{formula}{formula block}

\item{rateName}{string name to identify the rate}
}
\description{
Function to parse formula when there is no model
}
