% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parsing.R
\name{separate_name_and_constraint}
\alias{separate_name_and_constraint}
\title{Separate the name of the variable from the constraint}
\usage{
separate_name_and_constraint(name_with_constraint)
}
\arguments{
\item{name_with_constraint}{a name and constraint text chunk}
}
\description{
Separate the name of the variable from the constraint
}
