% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model.R
\name{transition}
\alias{transition}
\title{transition
constructs and validates an object of class \code{transition}}
\usage{
transition(name = NULL, parent, offspring, model = NULL)
}
\arguments{
\item{name}{a name for the transition event}

\item{parent}{the parent type for the transition}

\item{offspring}{the update to the system after the transition}

\item{model}{the user-defined bp_model output of process_model}
}
\value{
The \code{transition} object if it is valid, throws an error otherwise
}
\description{
transition
constructs and validates an object of class \code{transition}
}
