% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model.R
\name{validate_process_model}
\alias{validate_process_model}
\title{validate_process_model
verifies the correctness of an object of class \code{bp_model}}
\usage{
validate_process_model(
  bp_model,
  parameter_constraints = NULL,
  priors = NULL,
  predictor_names = NULL,
  observation_error = FALSE,
  hierarchical = NULL
)
}
\value{
The \code{process_model} object if it is valid, throws an error otherwise
}
\description{
validate_process_model
verifies the correctness of an object of class \code{bp_model}
}
