% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model.R
\name{bp_model}
\alias{bp_model}
\title{constructor for object of class \code{bp_model}}
\usage{
bp_model(...)
}
\arguments{
\item{...}{a transition list structured as a vector of transitions}
}
\value{
A new branching process model object
}
\description{
constructor for object of class \code{bp_model}
}
