% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model.R
\name{process_model}
\alias{process_model}
\title{process_model
constructs and validates an object of class \code{process_model}}
\usage{
process_model(
  bp_model,
  parameter_constraints = NULL,
  priors = NULL,
  predictor_names = NULL,
  observation_error = FALSE,
  hierarchical = NULL
)
}
\arguments{
\item{bp_model}{a bp_model object containing user-defined transitions}

\item{priors}{vector of characters listing priors for any parameters}

\item{observation_error}{logical to include error due to observation}

\item{hierarchical}{vector of parameters that are random effects and have
hyperparameters}

\item{parameter_constriants}{vector of bounds on the parameters}
}
\value{
The \code{process_model} object if it is valid, throws an error otherwise
}
\description{
process_model
constructs and validates an object of class \code{process_model}
}
